refactor: split CourseBuilder into 18-file modular folder
Main CourseBuilder.jsx: 3,489 → 341 lines (imports only)
New folder: client/src/components/CourseBuilder/

Files:
- index.js (re-export entry point)
- CourseBuilder.jsx (main component, tabs, save/load)
- AIGenerator.jsx (AI course generation with Anthropic)
- ContentEditor.jsx (block canvas with module sidebar)
- BlockEditor.jsx (all 17 block type editors)
- BlockPicker.jsx (block type selector grid)
- BlockTypeCatalog.jsx (reference view of all types)
- ExamGenerator.jsx (AI exam question generation)
- ReferencesManager.jsx (APA reference management)
- ACEPChecker.jsx (NBCC ACEP compliance checker)
- ImportTab.jsx (docx/pdf/md importer)
- NarrationTab.jsx (audio narration wrapper)
- NarrationPanel.jsx (narration controls)
- CloudinaryUploader.jsx (image upload)
- InsertBar.jsx (between-block insert UI)
- constants.js (colors, block types, ACEP rules)
- styles.js (shared style objects)
- utils.js (countWords, uid, countBlockWords)

App.jsx import unchanged: ./components/CourseBuilder → folder index.js
